# packages/api-client/src/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CTW 2.0 API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/deals": {
      "get": {
        "summary": "List deals",
        "tags": [
          "deals"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "organizationId": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "primaryCompanyName": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "stage": {
                        "type": "string",
                        "enum": [
                          "prospect",
                          "loi",
                          "negotiation",
                          "diligence",
                          "close",
                          "closed_won",
                          "lost"
                        ]
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "active",
                          "archived",
                          "closed"
                        ]
                      },
                      "ownerInitials": {
                        "type": "string"
                      },
                      "staleFlag": {
                        "type": "boolean"
                      },
                      "lastActivityAt": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "nextActionLabel": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "pendingApprovals": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 9007199254740991
                      },
                      "capabilities": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    },
                    "required": [
                      "id",
                      "organizationId",
                      "title",
                      "primaryCompanyName",
                      "stage",
                      "status",
                      "ownerInitials",
                      "staleFlag",
                      "lastActivityAt",
                      "nextActionLabel",
                      "pendingApprovals",
                      "capabilities"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create deal",
        "tags": [
          "deals"
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "primaryCompanyName": {
                    "type": "string"
                  }
                },
                "required": [
                  "title"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "organizationId": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "primaryCompanyName": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "stage": {
                      "type": "string",
                      "enum": [
                        "prospect",
                        "loi",
                        "negotiation",
                        "diligence",
                        "close",
                        "closed_won",
                        "lost"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "active",
                        "archived",
                        "closed"
                      ]
                    },
                    "ownerInitials": {
                      "type": "string"
                    },
                    "staleFlag": {
                      "type": "boolean"
                    },
                    "lastActivityAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "nextActionLabel": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "pendingApprovals": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "capabilities": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "id",
                    "organizationId",
                    "title",
                    "primaryCompanyName",
                    "stage",
                    "status",
                    "ownerInitials",
                    "staleFlag",
                    "lastActivityAt",
                    "nextActionLabel",
                    "pendingApprovals",
                    "capabilities"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/deals/{dealId}": {
      "get": {
        "summary": "Get deal",
        "tags": [
          "deals"
        ],
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "organizationId": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "primaryCompanyName": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "stage": {
                      "type": "string",
                      "enum": [
                        "prospect",
                        "loi",
                        "negotiation",
                        "diligence",
                        "close",
                        "closed_won",
                        "lost"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "active",
                        "archived",
                        "closed"
                      ]
                    },
                    "ownerInitials": {
                      "type": "string"
                    },
                    "staleFlag": {
                      "type": "boolean"
                    },
                    "lastActivityAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "nextActionLabel": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "pendingApprovals": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "capabilities": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "id",
                    "organizationId",
                    "title",
                    "primaryCompanyName",
                    "stage",
                    "status",
                    "ownerInitials",
                    "staleFlag",
                    "lastActivityAt",
                    "nextActionLabel",
                    "pendingApprovals",
                    "capabilities"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch deal",
        "tags": [
          "deals"
        ],
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1
                  },
                  "primaryCompanyName": {
                    "type": "string"
                  },
                  "staleFlag": {
                    "type": "boolean"
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "organizationId": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "primaryCompanyName": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "stage": {
                      "type": "string",
                      "enum": [
                        "prospect",
                        "loi",
                        "negotiation",
                        "diligence",
                        "close",
                        "closed_won",
                        "lost"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "active",
                        "archived",
                        "closed"
                      ]
                    },
                    "ownerInitials": {
                      "type": "string"
                    },
                    "staleFlag": {
                      "type": "boolean"
                    },
                    "lastActivityAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "nextActionLabel": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "pendingApprovals": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "capabilities": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "id",
                    "organizationId",
                    "title",
                    "primaryCompanyName",
                    "stage",
                    "status",
                    "ownerInitials",
                    "staleFlag",
                    "lastActivityAt",
                    "nextActionLabel",
                    "pendingApprovals",
                    "capabilities"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/deals/{dealId}/move-stage": {
      "post": {
        "summary": "Move deal stage",
        "tags": [
          "deals"
        ],
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "stage": {
                    "type": "string",
                    "enum": [
                      "prospect",
                      "loi",
                      "negotiation",
                      "diligence",
                      "close",
                      "closed_won",
                      "lost"
                    ]
                  }
                },
                "required": [
                  "stage"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "organizationId": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "primaryCompanyName": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "stage": {
                      "type": "string",
                      "enum": [
                        "prospect",
                        "loi",
                        "negotiation",
                        "diligence",
                        "close",
                        "closed_won",
                        "lost"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "active",
                        "archived",
                        "closed"
                      ]
                    },
                    "ownerInitials": {
                      "type": "string"
                    },
                    "staleFlag": {
                      "type": "boolean"
                    },
                    "lastActivityAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "nextActionLabel": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "pendingApprovals": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "capabilities": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "id",
                    "organizationId",
                    "title",
                    "primaryCompanyName",
                    "stage",
                    "status",
                    "ownerInitials",
                    "staleFlag",
                    "lastActivityAt",
                    "nextActionLabel",
                    "pendingApprovals",
                    "capabilities"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/deals/{dealId}/archive": {
      "post": {
        "summary": "Archive deal",
        "tags": [
          "deals"
        ],
        "parameters": [
          {
            "name": "dealId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "organizationId": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "primaryCompanyName": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "stage": {
                      "type": "string",
                      "enum": [
                        "prospect",
                        "loi",
                        "negotiation",
                        "diligence",
                        "close",
                        "closed_won",
                        "lost"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "active",
                        "archived",
                        "closed"
                      ]
                    },
                    "ownerInitials": {
                      "type": "string"
                    },
                    "staleFlag": {
                      "type": "boolean"
                    },
                    "lastActivityAt": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "nextActionLabel": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "pendingApprovals": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "capabilities": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "id",
                    "organizationId",
                    "title",
                    "primaryCompanyName",
                    "stage",
                    "status",
                    "ownerInitials",
                    "staleFlag",
                    "lastActivityAt",
                    "nextActionLabel",
                    "pendingApprovals",
                    "capabilities"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
  }
}